telegram-weather-bot/requirements.txt
aiogram>=3.13.0
aiohttp>=3.10.0
python-dotenv>=1.0.0
